Feature/google-analytics
Fix method sendMetrics
